pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page_analyzer"
authors = ["leilanimoone <leilanimoone@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
